Update 4
To the basic components program: Added a component that use this. to display a name variable.  Added a component that adds an event handler to a button that displays a string message to the screen on button click.  Added new jsx and index html file for component interactions experimentation.  Added a component to this file that takes a message input into the PropsDisplayer component and turns its props object into a string then displays the props to the screen in an h2 tag.