Update 5
Added divs to the index file for react components to be appended to.  Added Greeting component that takes an input property an displays it within its JSX.  Added two components: AnotherGreeting and App.  App passes properties to AnotherGreeting and displays them below an h1 when rendered.  Added two components Button and Talker.  Talker has a method that plays an alert.  Talker passes its talk method to Button component.  Button includes a button element which includes onClick event-handler which calls Talker's talk method.  Added two components List and ChildrenApp.  ChildrenApp calls two instances of the List component.  List takes the input properties and displays the children as an h1 for the title and ul element for the children.