Update 10
Changed the friends array to contain multiple objects with a name and active status. Added active friends component.  The component maps all friends with active status of true and renders them with a button to remove the friend or change their status to deactive on click which moves them to the deactive list.  The deactive component has similar functionality except maps the deactive friends and a button for each friend that sends the friend back to active status.  Created a method in the app component that is used for designating whether a friend is active or inactive.