Update 12
Created new folder for api program.  Added boilerplate code including a window.API which holds a promise and returns the top repos as data from the github api.  Added Nav component that holds the languages in an an array and maps them to the screen.  Added state to the app component that will hold the repos, sets active language and a loading property.  Added fetchRepos method in the app component that gets the promise from the window.API and updates the App components repos property with the repo data and sets loading state to false.  Added functionality to render function to display Nav component and if loading state isn't tru to display the data else to display loading component.